+++ src/AtmosphereModels/AtmosphereModels.jl
@@ -1,5 +1,6 @@
 module AtmosphereModels
 
 include("atmosphere_model.jl")
+include("update_atmosphere_model_state.jl")
 
 end
