+++ src/MoistThermodynamics/atmosphere_thermodynamics.jl
@@ -144,56 +144,17 @@ function AtmosphereThermodynamics(FT = Oceananigans.defaults.FloatType;
                                     dry_air, vapor, saturation, condensation, deposition)
 end
 
-#####
-##### Computations
-#####
-
 const AT = AtmosphereThermodynamics
 const IG = IdealGas
 
 @inline vapor_gas_constant(thermo::AT)    = thermo.molar_gas_constant / thermo.vapor.molar_mass
 @inline dry_air_gas_constant(thermo::AT)  = thermo.molar_gas_constant / thermo.dry_air.molar_mass
 
-const NonCondensingAtmosphereThermodynamics{FT} = AtmosphereThermodynamics{FT, Nothing} 
-
-"""
-    saturation_vapor_pressure(T, thermo)
-
-Compute the saturation vapor pressure over a liquid surface by integrating
-the Clausius-Clapeyron relation,
-
-```math
-dp/dT = ℒᵛ / (Rᵛ T^2)
-```
-"""
-@inline function saturation_vapor_pressure(T, thermo, phase_transition::PhaseTransition)
-    ℒ₀  = phase_transition.latent_heat
-    cᵖˡ = phase_transition.heat_capacity
-    T₀  = thermo.saturation.energy_reference_temperature
-    Tᵗʳ = thermo.saturation.triple_point_temperature
-    pᵗʳ = thermo.saturation.triple_point_pressure
-    cᵖᵛ = thermo.vapor.heat_capacity
-    Rᵛ  = vapor_gas_constant(thermo)
-
-    Δcᵖ = cᵖˡ - cᵖᵛ
-    Δϰ = Δcᵖ / Rᵛ
-
-    return pᵗʳ * (T / Tᵗʳ)^Δϰ * exp((ℒ₀ - Δcᵖ * T₀) * (1/Tᵗʳ - 1/T) / Rᵛ)
-end
-
-# Over a liquid surface
-@inline function saturation_specific_humidity(T, ρ, thermo, phase_transition::PhaseTransition)
-    p★ = saturation_vapor_pressure(T, thermo, phase_transition)
-    Rᵛ = vapor_gas_constant(thermo)
-    return p★ / (ρ * Rᵛ * T)
-end
-
 @inline function mixture_gas_constant(q, thermo::AT)
     Rᵈ = dry_air_gas_constant(thermo)   
     Rᵛ = vapor_gas_constant(thermo)   
     return Rᵈ * (1 - q) + Rᵛ * q
 end
-
 
 """
     mixture_heat_capacity(q, thermo)
@@ -206,107 +167,4 @@ liquid condensate.
     cᵖᵈ = thermo.dry_air.heat_capacity
     cᵖᵛ = thermo.vapor.heat_capacity
     return cᵖᵈ * (1 - q) + cᵖᵛ * q
-end
-
-#=
-For example, if we would like to account for the mass of condensate consistently:
-struct WarmCondensate{FT}
-    vapor :: FT
-    liquid :: FT
-    total :: FT
-end
-
-# Then more correctly:
-@inline function mixture_heat_capacity(q::WarmCondensate, thermo::AT)
-    cᵖᵈ = dry_air_heat_capacity(thermo)
-    cᵖᵛ = vapor_heat_capacity(thermo)
-    cᵖˡ = thermo.condensation.liquid_heat_capacity
-    qᵗ = q.total
-    qᵛ = q.vapor
-    qˡ = q.liquid
-    return cᵖᵈ * (1 - qᵗ) + cᵖᵛ * qᵛ + cᵖˡ * qˡ
-end
-
-@inline function mixture_gas_constant(q::WarmCondensate, thermo::AT)
-=#
-
-#####
-##### state thermodynamics for a Boussinesq model
-#####
-
-# Organizing information about the state is a WIP
-struct ThermodynamicState{FT}
-    θ :: FT
-    q :: FT
-    z :: FT
-end
-
-struct ReferenceState{FT}
-    p₀ :: FT # base pressure: reference pressure at z=0
-    θ :: FT  # constant reference potential temperature
-end
-
-function ReferenceState(FT = Oceananigans.defaults.FloatType;
-                        base_pressure = 101325,
-                        potential_temperature = 288)
-
-    return ReferenceState{FT}(convert(FT, base_pressure),
-                              convert(FT, potential_temperature))
-end
-
-"""
-    reference_density(ref, thermo)
-
-Compute the reference density associated with the reference pressure and potential temperature.
-The reference density is defined as the density of dry air at the reference pressure and temperature.
-"""
-@inline function reference_density(z, ref, thermo)
-    Rᵈ = dry_air_gas_constant(thermo)
-    p = reference_pressure(z, ref, thermo)
-    return p / (Rᵈ * ref.θ)
-end
-
-@inline function base_density(ref, thermo)
-    Rᵈ = dry_air_gas_constant(thermo)
-    return ref.p₀ / (Rᵈ * ref.θ)
-end
-
-@inline function reference_specific_volume(z, ref, thermo)
-    Rᵈ = dry_air_gas_constant(thermo)
-    p = reference_pressure(z, ref, thermo)
-    return Rᵈ * ref.θ / p
-end
-
-@inline function reference_pressure(z, ref, thermo)
-    cᵖᵈ = thermo.dry_air.heat_capacity
-    Rᵈ = dry_air_gas_constant(thermo)
-    ϰᵈ⁻¹ = Rᵈ / cᵖᵈ
-    g = thermo.gravitational_acceleration
-    return ref.p₀ * (1 - g * z / (cᵖᵈ * ref.θ))^ϰᵈ⁻¹
-end
-
-@inline function saturation_specific_humidity(T, z, ref::ReferenceState, thermo, phase_transition)
-    ρ = reference_density(z, ref, thermo)
-    return saturation_specific_humidity(T, ρ, thermo, phase_transition)
-end
-
-@inline function exner_function(state, ref, thermo)
-    Rᵐ = mixture_gas_constant(state.q, thermo)
-    cᵖᵐ = mixture_heat_capacity(state.q, thermo)
-    inv_ϰᵐ = Rᵐ / cᵖᵐ
-    pᵣ = reference_pressure(state.z, ref, thermo)
-    return (pᵣ / ref.p₀)^inv_ϰᵐ
-end
-
-condensate_specific_humidity(T, state, ref, thermo) =
-    condensate_specific_humidity(T, state.q, state.z, ref, thermo)
-
-function condensate_specific_humidity(T, q, z, ref, thermo)
-    qᵛ★ = saturation_specific_humidity(T, z, ref, thermo, thermo.condensation)
-    return max(0, q - qᵛ★)
-end
-
-function ice_specific_humidity(T, q, z, ref, thermo)
-    qi★ = saturation_specific_humidity(T, z, ref, thermo, thermo.deposition)
-    return max(0, q - qi★)
 end
